@@ -17,17 +17,17 @@
     "audit:check": "yarn audit --json --production > yarn-audit-current || true && jq -s '[.[] | select(.type==\"auditAdvisory\") | .data.advisory.id] | sort' yarn-audit-current > current-ids.json && jq -s '[.[] | select(.type==\"auditAdvisory\") | .data.advisory.id] | sort' yarn-audit-known-issues > known-ids.json && diff -q known-ids.json current-ids.json || (echo '❌ New vulnerabilities detected. Please review.' && exit 1)"
   },
   "dependencies": {
-    "@angular/animations": "^19.2.4",
-    "@angular/common": "^19.2.4",
-    "@angular/compiler": "^19.2.4",
-    "@angular/core": "^19.2.4",
-    "@angular/forms": "^19.2.4",
-    "@angular/platform-browser": "^19.2.4",
-    "@angular/platform-browser-dynamic": "^19.2.4",
-    "@angular/router": "^19.2.4",
+    "@angular/animations": "^20.1.4",
+    "@angular/common": "^20.1.4",
+    "@angular/compiler": "^20.1.4",
+    "@angular/core": "^20.1.4",
+    "@angular/forms": "^20.1.4",
+    "@angular/platform-browser": "^20.1.4",
+    "@angular/platform-browser-dynamic": "^20.1.4",
+    "@angular/router": "^20.1.4",
     "@microsoft/applicationinsights-web": "^3.3.6",
     "@ministryofjustice/frontend": "^5.0.0",
-    "@ngrx/signals": "^19.0.1",
+    "@ngrx/signals": "^20.0.0",
     "accessible-autocomplete": "^3.0.1",
     "fast-sort": "^3.4.1",
     "govuk-frontend": "^5.9.0",
@@ -40,12 +40,12 @@
     "zone.js": "~0.15.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^19.2.5",
-    "@angular/cli": "^19.2.5",
-    "@angular/compiler-cli": "^19.2.4",
+    "@angular-devkit/build-angular": "^20.0.0",
+    "@angular/cli": "^20.0.0",
+    "@angular/compiler-cli": "^20.1.4",
     "@types/jasmine": "~5.1.0",
     "@types/luxon": "^3.4.2",
-    "angular-eslint": "19.8.1",
+    "angular-eslint": "20.1.1",
     "eslint": "^9.21.0",
     "jasmine-core": "~5.9.0",
     "karma": "~6.4.0",
