@@ -1,4 +1,4 @@
-import { Pipe, PipeTransform } from '@angular/core';
+import { inject, Pipe, PipeTransform } from '@angular/core';
 import { DateService } from '@hmcts/opal-frontend-common/services/date-service';
 
 @Pipe({
@@ -6,7 +6,7 @@ import { DateService } from '@hmcts/opal-frontend-common/services/date-service';
   standalone: true,
 })
 export class DateFormatPipe implements PipeTransform {
-  constructor(private readonly dateService: DateService) {}
+  private readonly dateService = inject(DateService);
 
   transform(value: string | null | undefined, inputFormat: string, outputFormat: string): string {
     if (!value) return '—';
