@@ -1,6 +1,7 @@
 import { TestBed } from '@angular/core/testing';
 import { TransformationService } from './transformation.service';
 import { ITransformItem } from '@hmcts/opal-frontend-common/services/transformation-service/interfaces';
+import { TRANSFORM_ITEM_DEFAULTS } from './constants/transform-item.constants';
 
 describe('TransformationService', () => {
   let service: TransformationService;
@@ -21,10 +22,24 @@ describe('TransformationService', () => {
     it('should return the original value if no transformation is applied', () => {
       const value = 'test';
       const transformItem: ITransformItem = {
+        ...TRANSFORM_ITEM_DEFAULTS,
         key: 'testKey',
         transformType: 'none',
-        dateInputFormat: null,
-        dateOutputFormat: null,
+      };
+      const result = service['applyTransformation'](value, transformItem);
+      expect(result).toBe(value);
+    });
+
+    it('should return the original value if no value is sent', () => {
+      const value = null;
+      const transformItem: ITransformItem = {
+        ...TRANSFORM_ITEM_DEFAULTS,
+        key: 'testKey',
+        transformType: 'date',
+        dateConfig: {
+          inputFormat: 'dd/MM/yyyy',
+          outputFormat: 'yyyy-MM-dd',
+        },
       };
       const result = service['applyTransformation'](value, transformItem);
       expect(result).toBe(value);
@@ -33,14 +48,65 @@ describe('TransformationService', () => {
     it('should transform date values correctly', () => {
       const value = '04/06/1991';
       const transformItem: ITransformItem = {
+        ...TRANSFORM_ITEM_DEFAULTS,
         key: 'dateKey',
         transformType: 'date',
-        dateInputFormat: 'dd/MM/yyyy',
-        dateOutputFormat: 'yyyy-MM-dd',
+        dateConfig: {
+          inputFormat: 'dd/MM/yyyy',
+          outputFormat: 'yyyy-MM-dd',
+        },
       };
 
       const result = service['applyTransformation'](value, transformItem);
       expect(result).toBe('1991-06-04');
+    });
+
+    it('should not transform date if date is not valid', () => {
+      const value = '13/13/1991';
+      const transformItem: ITransformItem = {
+        ...TRANSFORM_ITEM_DEFAULTS,
+        key: 'dateKey',
+        transformType: 'date',
+        dateConfig: {
+          inputFormat: 'dd/MM/yyyy',
+          outputFormat: 'yyyy-MM-dd',
+        },
+      };
+
+      const result = service['applyTransformation'](value, transformItem);
+      expect(result).toBe('13/13/1991');
+    });
+
+    it('should transform time values correctly by adding offset', () => {
+      const value = '14:30';
+      const transformItem: ITransformItem = {
+        ...TRANSFORM_ITEM_DEFAULTS,
+        key: 'timeKey',
+        transformType: 'time',
+        timeConfig: {
+          addOffset: true,
+          removeOffset: false,
+        },
+      };
+
+      const result = service['applyTransformation'](value, transformItem);
+      expect(result).toBe('14:30:00Z');
+    });
+
+    it('should transform time values correctly by removing offset', () => {
+      const value = '14:30:00Z';
+      const transformItem: ITransformItem = {
+        ...TRANSFORM_ITEM_DEFAULTS,
+        key: 'timeKey',
+        transformType: 'time',
+        timeConfig: {
+          removeOffset: true,
+          addOffset: false,
+        },
+      };
+
+      const result = service['applyTransformation'](value, transformItem);
+      expect(result).toBe('14:30');
     });
   });
 
@@ -56,10 +122,13 @@ describe('TransformationService', () => {
       const input = { dateKey: '01/01/1991', otherKey: 'value' };
       const transformItems: ITransformItem[] = [
         {
+          ...TRANSFORM_ITEM_DEFAULTS,
           key: 'dateKey',
           transformType: 'date',
-          dateInputFormat: 'dd/MM/yyyy',
-          dateOutputFormat: 'yyyy-MM-dd',
+          dateConfig: {
+            inputFormat: 'dd/MM/yyyy',
+            outputFormat: 'yyyy-MM-dd',
+          },
         },
       ];
 
@@ -71,13 +140,32 @@ describe('TransformationService', () => {
       const input = { nested: { dateKey: '01/01/1991' } };
       const transformItems: ITransformItem[] = [
         {
+          ...TRANSFORM_ITEM_DEFAULTS,
           key: 'dateKey',
           transformType: 'date',
-          dateInputFormat: 'dd/MM/yyyy',
-          dateOutputFormat: 'yyyy-MM-dd',
+          dateConfig: {
+            inputFormat: 'dd/MM/yyyy',
+            outputFormat: 'yyyy-MM-dd',
+          },
         },
       ];
       expect(service.transformObjectValues(input, transformItems)).toEqual({ nested: { dateKey: '1991-01-01' } });
+    });
+
+    it('should transform objects in arrays', () => {
+      const input = { nested: [{ dateKey: '01/01/1991' }] };
+      const transformItems: ITransformItem[] = [
+        {
+          ...TRANSFORM_ITEM_DEFAULTS,
+          key: 'dateKey',
+          transformType: 'date',
+          dateConfig: {
+            inputFormat: 'dd/MM/yyyy',
+            outputFormat: 'yyyy-MM-dd',
+          },
+        },
+      ];
+      expect(service.transformObjectValues(input, transformItems)).toEqual({ nested: [{ dateKey: '1991-01-01' }] });
     });
   });
 
